feat: parse double to float in sale_listing_signed_contract (#13252)

+++ dags/rede/reverse_listings_report_load/queries/reverse/sale_listing_signed_contract.sql
@@ -4,7 +4,7 @@ SELECT
     fsde.sk_house AS property_id,
     dc.uuid_company AS company_uuid,
     'SALE' AS business_context,
-    dsa.sale_price_agreed AS contract_value,
+    dsa.sale_price_agreed::FLOAT AS contract_value,
     fsde.ts_event,
     fsde.year,
     fsde.month,
